requirements.txt
pytest==8.0.0
selenium==4.18.1
pytest-xdist==3.5.0
allure-pytest==2.13.5
webdriver_manager==4.0.1